src/utils/random_utils.py
import torch
import numpy as np
import random
import os
from src.utils.io_utils import get_rank

def set_random_seed(seed):
    seed = seed + get_rank()
    random.seed(seed)
    os.environ['PYTHONHASHSEED'] = str(seed)
    np.random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)
    torch.backends.cudnn.benchmark = False
    torch.backends.cudnn.deterministic = True